+++ Java/ds/maxPriorityQueue.java
@@ -1,35 +1,23 @@
-//Implementing the class for Max Priority Queue which includes the required functions....
-
-
+//Implementing the class for Max Priority Queue which includes the required functions......
+//Max Priority Queue is a data structure which manage a list of values, it gives priority to element with maximum value i.e elements are inserted in the order 
+//in which they arrive the queue and the maximum value is always removed first from the queue.
 import java.util.ArrayList;
 import java.util.Scanner;
-public class MaxPriorityQueue {
+public class maxPriorityQueue {
 	private ArrayList<Integer> heap=new ArrayList<>();
-
-    // public  PriorityQueue() {
-    //     heap = new ArrayList<Integer>();
-    // }
-	// Complete this class
 	boolean isEmpty() {
 	 return heap.size() == 0;
-        // Implement the isEmpty() function here
 	}
-
 	int getSize() {
-		// Implement the getSize() function here
          return heap.size();
 	}
-
 	int getMax() {
-		// Implement the getMax() function here
           if(isEmpty())
             return Integer.MIN_VALUE;
         return heap.get(0);
 	}
-
 	void insert(int element) {
-		// Implement the insert() function here
-          heap.add(element);
+         heap.add(element);
         int childIndex = heap.size() - 1;
         int parentIndex = (childIndex - 1) / 2;
         while(childIndex > 0)
@@ -41,70 +29,55 @@ public class MaxPriorityQueue {
                 heap.set(parentIndex, temp);
                 childIndex = parentIndex;
                 parentIndex = (childIndex - 1) / 2;
-            }
-            else
+            }else{
                 return;
-        }
-	}
-
-	int removeMax() {
-		// Implement the removeMax() function here
+           }
+	    }	
+     }
+        int removeMax() {
         if(isEmpty())
             return Integer.MIN_VALUE;
-
         int remove=heap.get(0);
         heap.set(0,heap.get(heap.size()-1));
         heap.remove(heap.size()-1);
-
         int parentIndex = 0;
         int leftchildIndex = (2*parentIndex)+1;
         int rightchildIndex = (2*parentIndex)+2;
-
         while(leftchildIndex<heap.size())
         {
             int p=heap.get(parentIndex);
             int l=heap.get(leftchildIndex);
-            if(rightchildIndex>=heap.size())
-            {
-                if(p<l)
-                {
+            if(rightchildIndex>=heap.size()){
+                if(p<l){
                     heap.set(parentIndex,l);
                     heap.set(leftchildIndex,p);
                     return remove;
-                }
-                else
+                }else{
                     return remove;
+		        }
             }
             int r=heap.get(rightchildIndex);
-            if(p>l&&p>r)
+             if(p>l&&p>r){
                 return remove;
-            else if(p<l&&p<r)
-            {
-                if(l>=r)
-                {
-                    heap.set(parentIndex,l);
+	         }else if(p<l&&p<r){
+                if(l>=r){
+                   heap.set(parentIndex,l);
                     heap.set(leftchildIndex,p);
                     parentIndex =leftchildIndex;
-                }
-                else
-                {
+                }else{
                     heap.set(parentIndex,r);
                     heap.set(rightchildIndex,p);
                     parentIndex =rightchildIndex;
                 }
                 leftchildIndex = (2*parentIndex)+1;
                 rightchildIndex = (2*parentIndex)+2;
-            }
-            else if(p<l)
-            {
+            }else if(p<l){
                 heap.set(parentIndex,l);
                 heap.set(leftchildIndex,p);
                 parentIndex =leftchildIndex;
                 leftchildIndex = (2*parentIndex)+1;
                 rightchildIndex = (2*parentIndex)+2;
-            }
-            else
-            {
+            }else{
                 heap.set(parentIndex,r);
                 heap.set(rightchildIndex,p);
                 parentIndex =rightchildIndex;
@@ -115,3 +88,9 @@ public class MaxPriorityQueue {
         return remove;
 	}
 }
+/*
+ Sample Input
+ 6 4 2 12 10 
+ Sample Output
+ 12 10 8 6 4 2
+*/
